requirements.txt
# Core runtime dependencies (Python 3.11)
# Pin ranges chosen for stability and widewheel availability on Linux/WSL2

# Broker & connectivity
ib-insync>=0.9.86,<1.0.0

# Config & CLI
pydantic>=2.5,<3
python-dotenv>=1.0,<2
PyYAML>=6.0.1,<7
typer>=0.9,<1.0

# Data & computing
numpy>=1.26
pandas>=2.2,<3
pyarrow>=15,<18
pandas-market-calendars>=4,<6
requests>=2.31,<3

# Indicators & analytics
ta>=0.11,<0.12

# Reporting
plotly>=5.20,<6
Jinja2>=3.1,<4

# Historical data fallback for corporate actions factors (optional but recommended)
yfinance>=0.2.30,<0.3

# --- Development & QA (you can move these to requirements-dev.txt if preferred) ---
pytest>=7.4,<9
pytest-asyncio>=0.21,<0.24
mypy>=1.8,<2
ruff>=0.3,<0.6
black>=24.3,<25
pre-commit>=3.6,<4
ipykernel>=6.29,<7
jupyterlab>=4.1,<5
